Rename: Auth 페이지 내 컴포넌트 이동
